pyproject.toml
[tool.poetry]
name = "rsna2024"
version = "0.1.0"
description = ""
authors = ["Daniel Phalen <phalendj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
torchio = "^0.19.9"
torch = "^2.4.0"
numpy = "^2.0.1"
pandas = "^2.2.2"
pytest = "^8.3.2"
scikit-learn = "^1.5.1"
tqdm = "^4.66.4"
pydicom = "^2.4.4"
opencv-python = "^4.10.0.84"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
timm = "^1.0.8"
transformers = "^4.43.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
